+++ app/controllers/pages_controller.rb
@@ -1,11 +1,4 @@
 class PagesController < ApplicationController
   def home
   end
-
-  def up
-    Redis.current.ping
-    ActiveRecord::Base.connection.execute("SELECT 1")
-
-    head :ok
-  end
 end