+++ config/routes.rb
@@ -1,9 +1,8 @@
 Rails.application.routes.draw do
   root to: "pages#home"
 
-  scope controller: :pages do
-    get :up
-  end
+  get "/up/", to: "up#index", as: :up
+  get "/up/databases", to: "up#databases", as: :up_databases
 
   # Sidekiq has a web dashboard which you can enable below. It's turned off by
   # default because you very likely wouldn't want this to be available to